+++ 1D/Population.cpp
@@ -17,16 +17,16 @@ double a0 = 10, aex1 = 20, aex2 = 20, sigmab0 = .5, sigmabex1 = .1, sigmabex2 = 
 double g(double x, double sigma) { return exp(-x * x / (2 * sigma * sigma)); }
 double b(double x) { return a0 * g(x, sigmab0) + aex1 * g(x - 2, sigmabex1) + aex2 * g(x + 2, sigmabex2) + slope * x; }
 double c(double x, double y) { return .1 * max(-Ac * (x - y) * (x - y) * (x - y) * (x - y) + (x - y) * (x - y) + 3, 0.0); }
-double rescaling(double x, int size) { return 6.0 * (x - 1.0) / (size - 1.0) - 3.0; }
+double rescaling(double x, size_t size) { return 6.0 * (x - 1.0) / (size - 1.0) - 3.0; }
 
-double NewNu(vector<double> *nup, double K, double p, int size) {
+double NewNu(vector<double> *nup, double K, double p, size_t size) {
     vector<double> &nu = *nup;
     vector<double>  B(size + 1, 0), De(size + 1, 0), M(size + 1, 0);
     double          BB = 0, DD = 0, MM = 0;
 
-    for (int x = 1; x <= size; x++) {
+    for (size_t x = 1; x <= size; x++) {
         double s = 0;
-        for (int y = 1; y <= size; ++y) { s += c(rescaling(x, size), rescaling(y, size)) * nu[y]; }
+        for (size_t y = 1; y <= size; ++y) { s += c(rescaling(x, size), rescaling(y, size)) * nu[y]; }
         B[x] = nu[x] * (1 - p) * b(rescaling(x, size));
         BB += B[x];
         M[x] = nu[x] * p * b(rescaling(x, size));
@@ -37,16 +37,16 @@ double NewNu(vector<double> *nup, double K, double p, int size) {
 
     double u = prng.uniform_real() * (BB + DD + MM);
     if (u < BB) {
-        int64_t z = 0;
-        double  v = prng.uniform_real() * BB;
+        size_t z = 0;
+        double v = prng.uniform_real() * BB;
         while (v > 0) {
             z++;
             v -= B[z];
         }
         nu[z] += 1 / K;
     } else if (u < BB + DD) {
-        int64_t z = 0;
-        double  v = prng.uniform_real() * DD;
+        size_t z = 0;
+        double v = prng.uniform_real() * DD;
         while (v > 0) {
             z++;
             v -= De[z];
@@ -54,14 +54,14 @@ double NewNu(vector<double> *nup, double K, double p, int size) {
         nu[z] -= 1 / K;
         if (nu[z] < .5 / K) nu[z] = 0;
     } else {
-        int64_t z = 0;
-        double  v = prng.uniform_real() * MM;
+        size_t z = 0;
+        double v = prng.uniform_real() * MM;
         while (v > 0) {
             z++;
             v -= M[z];
         }
         int64_t dz = prng.bernoulli() ? 1 : -1;
-        if ((z > 0) && (z + dz >= 1) && (z + dz <= size)) nu[z + dz] += 1 / K;
+        if ((z > 0) && (int(z) + dz >= 1) && (size_t(int(z) + dz) <= size)) nu[size_t(int(z) + dz)] += 1 / K;
     }
 
     return prng.exponential(K * (BB + DD + MM));
@@ -70,15 +70,15 @@ double NewNu(vector<double> *nup, double K, double p, int size) {
 int main(int argc, char **argv) {
     // n impair ! the function rescaling maps {1, ..., size} to {-3, ..., 3}; 7+(2^k*)6 is optimal rescaling
     vb::Hub H("Population dynamics", argc, argv, "s=4,K=100,p=.05,n=7,i=1000000");
-    prng.seed(int(H['s']));
+    prng.seed(size_t(H['s']));
     double K = H['K'], p = H['p'];
-    int    size = H['n'], shift = (size + 1) / 2, iterations = H['i'];
+    size_t size = H['n'], shift = (size + 1) / 2, iterations = H['i'];
 
     vector<double> nu(size + 1, 0);
     nu[shift] = 3;
 
     vector<cpx> pos(size, 0);
-    for (int64_t i = 0; i < size; ++i) pos[i] = cpx(i, nu[i + 1]);
+    for (size_t i = 0; i < size; ++i) pos[i] = cpx(i, nu[i + 1]);
 
     vb::Console W;
     W.manage(slope, -100.0, 100.0, "slope");
@@ -97,22 +97,22 @@ int main(int argc, char **argv) {
     vb::ProgressBar PB(iterations);
     double          t = 0;
 
-    for (int k = 1; k <= iterations; ++k) {
+    for (size_t k = 1; k <= iterations; ++k) {
         PB.set(k);
         t += NewNu(&nu, K, p, size);
         if (k % (iterations / 1000) == 0) {
             cout << k << " " << t;
-            for (int i = 1; i <= size; ++i) cout << " " << int(.2 + nu[i] * K);
+            for (size_t i = 1; i <= size; ++i) cout << " " << int(.2 + nu[i] * K);
             cout << endl;
             G.contents.clear();
-            for (int64_t i = 0; i < size; ++i) {
+            for (size_t i = 0; i < size; ++i) {
                 graph[i].emplace_back(1000.0 * k / iterations, nu[i + 1]);
-                G.add(std::make_unique<Path>(graph[i], vb::Pen(vb::Indexed(i))));
+                G.add(std::make_unique<Path>(graph[i], vb::Pen(vb::Indexed(int(i)))));
             }
             if (!G.visible()) { G.show(); }
             G.step();
         }
-        for (int64_t i = 0; i < size; ++i) pos[i] = cpx(i, nu[i + 1]);
+        for (size_t i = 0; i < size; ++i) pos[i] = cpx(i, nu[i + 1]);
         F.contents.clear();
         F.add(std::make_unique<Path>(pos));
         F.step();
